+++ src/types/zmodem-js.d.ts
@@ -1,4 +1,0 @@
-declare module 'zmodem.js/src/zmodem_browser.js' {
-  const zmodem: unknown
-  export default zmodem
-}